move pr labeler to separate repository
The YAML and Python script originated from the OMPI repository (diverged here:
https://github.com/open-mpi/ompi/commit/ef866a4657c782fef04513ee7334c7b762783b5e)

An OMPI maintainer wanted this and the git commit checker in two of their other
repos. Instead of copying the code to their repo(s), it is far simpler to move
the actions to separate repos and publish them on the GitHub marketplace.

The functionality remains the same, but the YAML had to be tweaked to work as a
reusable (marketplace) action. These changes are:

* add / change the action name, description, branding, and inputs
* remove the trigger
* remove permissions
* specify composite action (runs:)
* remove os specification
* remove code checkout (it wasn't necessary for the labeler in the first place)
* specify shell for each command run
* use github.action_path instead of the env variable GITHUB_WORKSPACE

Signed-off-by: Joe Downs <joe@dwns.dev>