github-labeler: check if the label exists before adding it
The GitHub API automatically creates a label when applying it (if it doesn't
exist already), so we need to check if the label exists in the repo before we go
to apply (and create) it. This way, the humans have control over which labels
are created and assigned, which is how the bot should work.